tdf#169318 tdf#127201 extension mgr: Open via new UNO command
Introduce a new UNO command, ".uno:ExtensionManager",
which is used to open the "Extensions" dialog from the
menu, notebookbar and using the keyboard shortcut
(Ctrl+Alt+E) instead of specifying the service name
"service:com.sun.star.deployment.ui.PackageManagerDialog"
in their configuration directly.

In a first step, handle the new UNO command in
SfxApplication::MiscExec_Impl by creating an
instance of the service by its name and executing the dialog
using the css::task::XJobExecutor interface without any
additional changes.
This matches what the previous way of triggering
the dialog would have done, i.e. no user-observable
behavior is intended yet.

This prepares for further changes in upcoming
commits, in particular passing a parent for
the dialog.

As documented in icon-themes/README.md, make sure to
have icons with the expected names, which is e.g.
used in the "Tabbed Compact" UI variant of the Notebookbar.
(Link to the already existing icons.)

Change-Id: I4b381b5f0b7cddd9f0aa9abd9b88c1c4364ab824
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193580
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>