tdf#169318 tdf#127201 extension mgr: Set dialog parent, make modal
So far, the "Extensions" dialog didn't have a parent set,
resulting in it not properly being shown on top of the
application window, while still blocking it when the
extension update dialog was started from the extensions
dialog.

Pass the dialog parent when creating it when handling the
UNO command, which fixes this problem when the dialog is
triggered via the UNO command, at least for the first run.

Also use the service ctor instead of creating the service
manually via the service name.

A remaining problem is that the parent for the
dialog in use is only set the first time the
dialog is opened, which can be seen e.g. as follows:

1) start Writer
2) "Tools" -> "Extensions"

-> Dialog opens modally on top of the Writer window
as expected

3) create another Writer doc
4) close the first Writer window
5) "Tools" -> "Extensions"

-> The dialog doesn't open modally on top of the
Writer window and the problem described in tdf#169318
is still reproducible.

Consistently setting the passed parent will have to be
addressed separately in the dialog implementation itself.

Running the dialog standalone via

    ./instdir/program/unopkg gui

still behaves as before.

Change-Id: Ieb7d7b564140b34ff5875d2e26a777b9932cfa2d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193585
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>