tdf#169318 extension mgr: Don't do too much when setting dlg title
In ServiceImpl::setDialogTitle, always remember the
explicitly set title and apply it in ServiceImpl::startExecuteModal
if non-empty (and keep remembering it), not only
when the dp_gui::TheExtensionManager::s_ExtMgr isn't set
yet.

Don't call TheExtensionManager::get (yet) which could
trigger more logic that seems more reasonable to do
later, when the dialog is actually run.
Simply skip setting the title at this point otherwise,
as it will be set in ServiceImpl::startExecuteModal
before the dialog gets shown anyway.

Change-Id: Id51627de09a5713996076eb0ca599cc361c54dd6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193587
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins