tdf#169320: Don't treat hidden-paragraph-mark as redline
The "hidden paragraph mark" feature is implemented to match the
feature of MS Word. In Writer, it is implemented using the same
"merged paragraphs" machinery as redlines; and some code doesn't
make distinction between the two.

The code, that calculates a number string, uses three internal
lists in SwTextNode to take redlines into account: mpNodeNum,
mpNodeNumRLHidden, mpNodeNumOrig, which reflect the hierarchy
effective for each mode. In "hidden redlines" mode, there are
no removed elements in the respective list. Therefore, number
string will be calculated from a different structure, and will
differ from the string generated from the original structure.
That is considered in SwTextFormatter::NewNumberPortion, which
calls SwTextNode::GetNumString with SwListRedlineType::HIDDEN
and SwListRedlineType::ORIGTEXT flags, compare the results, and
add the original text in [] after the current text.

Removal of elements hidden by redlines from mpNodeNumRLHidden
happens in CheckParaRedlineMerge, which finds fidden nodes, and
calls SwTextNode::RemoveFromListRLHidden for them. This change
makes sure that it is not called, when a node is not hidden by
a redline.

Change-Id: I93bc2c47bfefca115c1cbe47e53c7b0f7a119e4b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193583
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193603
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>