tdf#169318 extension mgr: Pass parent to TheExtensionManager::createDialog
So far, TheExtensionManager::m_xParent was holding
the parent window passed in the first call to
TheExtensionManager::get, and would use that when
creating a dialog in TheExtensionManager::createDialog.

Refactor this to instead pass the parent directly
to TheExtensionManager::createDialog and drop
TheExtensionManager::m_xParent.

This addresses the problem mentioned in earlier commit

    Change-Id: Ieb7d7b564140b34ff5875d2e26a777b9932cfa2d

    commit d5eb72a31ebb3687d9d53dab7e8128825a9bae8f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Nov 7 14:23:57 2025 +0100

        tdf#169318 tdf#127201 extension mgr: Set dialog parent, make modal

:

> A remaining problem is that the parent for the
> dialog in use is only set the first time the
> dialog is opened, which can be seen e.g. as follows:
>
> 1) start Writer
> 2) "Tools" -> "Extensions"
>
> -> Dialog opens modally on top of the Writer window
> as expected
>
> 3) create another Writer doc
> 4) close the first Writer window
> 5) "Tools" -> "Extensions"
>
> -> The dialog doesn't open modally on top of the
> Writer window and the problem described in tdf#169318
> is still reproducible.
>
> Consistently setting the passed parent will have to be
> addressed separately in the dialog implementation itself.

Now, the dialog opens properly modally on top of the
second Writer window in step 5.

With this commit in place, it's now also more obvious
that no parent is currently set for
UpdateRequiredDialog, as UpdateRequiredDialogService::execute
passes an empty Reference for the parent
(that was previously the default arg used for
TheExtensionManager::get).

Change-Id: Ifc06f093038812e262b5f35881701601b588bf19
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193617
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>