tdf#114441: Replaced sal_uLong in GDI MF code
The maximum size of the GDI metafile is 4GB, so
sal_uInt32 is sufficient for metafile size related
variables. The number of possible colors is 0xFFFFFF
so sal_uInt32 is sufficient to store the number of
colors. The R, G, B fields have a maximum of 255 so
sal_uInt8 is sufficient for storing them.

Change-Id: I0f0108df6eeb087ce1e799905c4f9ee03b409ab9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193608
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>