tdf#155855: fix pasting from CF_BITMAP
It has been implemented in commit e9391ae8eb3d0a4edf977f0159ee7e07a0a95859
(vcl113: #i111834# support pasting HBITMAP). But I fail to see how it
could ever work, despite the OOo bug was verified.

The problems were:

1. MIME type added for CF_BITMAP was exactly the same as for CF_DIB. Given
that the wanted flavor is passed to CDOTransferable::getTransferData in
DataFlavor struct, which uses the MIME type string, which is only turned
into CFormatEtc (encapsulating FORMATETC) in the function, it was never
possible to request CF_BITMAP: always CF_DIB got requested.

2. WinBITMAPToOOBMP used a call to GetBitmapDimensionEx to obtain the
dimensions out of HBITMAP. But that function is only used together with
SetBitmapDimensionEx; and the value that these two functions control is
documented to be only usable by applications, not by the system, and it
has nothing to do with the actual bitmap data stored in HBITMAP, it's a
kind of custom metadata. Generally, GetBitmapDimensionEx returns (0, 0)
for any bitmap in the clipboard. And then, it passed nullptr as HDC to
GetDIBits, which is 100% guaranteed to fail.

My theory is, that the CF_BITMAP-related code added in that patch always
was a dead code, and it happened to work because the system has a feature
of Synthesized Clipboard Formats, where CF_DIB will be created by system
on demand - so the change allowed to find and map CF_BITMAP in clipboard
to the MIME type shared with CF_DIB; then the MIME type was converted to
CF_DIB; and its functional code path was used.

However, this automatic generation of CF_DIB obviously doesn't happen in
case of Windows Clipboard History. Request for CF_DIB fails, so we have
to make the code for CF_BITMAP (the only data in the clipboard that the
history provides for image data) work.

This change adds a fallback to CDOTransferable::getClipboardData, that
tries CF_BITMAP when CF_DIB was requested and has failed. It also fixes
WinBITMAPToOOBMP.

Change-Id: I7d5eb64cc378510d303686d2bab9c53129607980
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193630
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
(cherry picked from commit 426a9af71d99c8610b28c3bf73fb8e3fd88e0a64)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193722
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>