Fix selecting other table when inserting relation with only 2 tables
In the dialog to insert a new relation, if there are only two tables
the list box to select the two tables tries to react to one of the
selected tables changing by making the other list box choose the other
table. However it was doing this by making the other list box flip its
selection. This breaks if the dialog is reacting to the user reselecting
the same table in a list box because the other list box flips its value
even though the value in the first list box hasn’t actually changed.

To fix the problem this patch makes it explicitly use the opposite table
from the first list box instead of just flipping its value.

Change-Id: Ibf5d39de9406c4617c33c8f81a8605cb61811f6e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193760
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>