Table style: fix wrong sidebar context after selecting
table style area. After we selected a draw object and then
we clicked inside a table style area the context was not updated
and only the cell context UI elements appeared.
Also need force the update in case of we set the cursor position
from one table cell area to another one, otherwise the sidebar
will not update the different Table area properties on the panel.

Change-Id: I70a25e30071779ad4053b967a3fb15ec0328d6cb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193690
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196758
Reviewed-by: Balazs Varga <balazs.varga@collabora.com>
Tested-by: Balazs Varga <balazs.varga@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/197984
Tested-by: Jenkins