Drop silencing clang-cl -Wextra-tokens in midl-generated code
The following commits introduced these:

77a39107b1d629022d3ba8711f7a7051e9461280 (Silence clang-cl -Wextra-tokens
in midl-generated code, 2015-11-28);
5ccdc4187bd679c41d89b920847fe0c04d81bdb8 (Silence clang-cl -Wextra-tokens
in midl-generated code, 2015-11-30);
34604446fc81be99ccee28f1977bf85e6abfab45 (Various clang-cl/loplugin
warnings, 2017-03-14).

The problem was, that some versions of midl compiler generated an invalid
code like "#endif !_MIDL_USE_GUIDDEF_", which caused warnings in gcc and
clang (see e.g. https://stackoverflow.com/q/29153518/1397376).

I see that as of VS 2022, this problem is not reproducible, so let's drop
these.

Change-Id: Icce797ae6fa8b1aca17ca0b5051e2af606126bf8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194010
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins