tdf#169316 sw markdown import: add a TemplateURL parameter
Writer markdown import creates a document with built-in Writer styles,
and it would be nice to instead use styles from a second template
document instead.

It is possible to use the File -> Templates functionality to manually
assign a template to the resulting document, then later update styles
from that, but it's far from straightforward when you want to simply do
a template + markdown -> PDF conversion.

Fix the problem by adding a new TemplateURL parameter for the Writer
markdown import: if that's provided, it can be a URL relative to the
markdown file and we'll load styles from that document before doing the
actual import. This is similar to how
SfxObjectShell::UpdateFromTemplate_Impl() works.

Example desktop cmdline:
soffice --infilter='Markdown:{"TemplateURL":{"type":"string","value":"./template.ott"}}' test.md

Change-Id: I1763d0e18a116056427a6d57bbefe3e29a8514d4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193715
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>