tdf#80959: fix percent character escaping in dp_misc::(en|de)codeString
The intention that these functions, that use % as a special character,
escape existing % with another %, was documented in comments at both
functions. But it had never been implemented properly.

In encodeString, the code did wrong calculations, and wrote first
special character (either %, or 0x00-0x0F) verbatim; and only then
processed following characters according to the algorithm. That meant
that an input with percent-encoded space, like

Foo%20Bar

was checked until the percent; then first four characters "Foo%" were
copied to the resulting buffer; and then the rest of characters were
sequentially checked for % or 0x00-0x0F (and none of them were). The
result was unchanged "Foo%20Bar", instead of expected "Foo%%20Bar".

In decodeString, despite the idea that %% must convert to %, the code
didn't expect anything except 0-9 and A-F after the first %. As the
result, an input string of "Foo%%20Bar" would become "Foo\EEBar", and
"Foo%20Bar" would become "Foo\020Bar".

This was that way since introduction of this code in AOO in commit
65415c716863d17196eae2a4a7f86f2e0ba5c1b6 (#i118662# remove usage of
BerkeleyDB in desktop module, 2011-12-12), ported to LO in commit
daeed90f4586eb9533041fb89bee163a5193596c (re-base on ALv2 code.
Includes:, 2012-11-19).

This change fixes and unifies the logic of both of the functions.

Change-Id: Ice5d078b007dbffacea247daaf81bde9c6386963
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194066
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins