tdf#169219 swpagerelsize ui: fix width percent value
This is a follow-up to 4.3 swpagerelsize

Without this patch, when you open up
sw/qa/extras/odfexport/data/relh-page-tdf80282.odt
and go to the image properties,
it would say relative width to entire page of 162 percent.
Now it says 100%.

Not sure why this wasn't handled earlier
when the stuff I copied and pasted from a few lines below
was initially added.
Perhaps because that situation wasn't really working at all
until my 6.1 commit which was flagged in this bug 169219.

Change-Id: I73163ad94bf895099576004f5d16242ffba5ca98
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193773
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins