tdf#89544: Display small, 3x3 pixel crosses for main grid points.
In Writer (and in Calc). To improve visibility of the grid.

The same happens in Draw and in Impress. Implement the same behaviour
for Witer (and Calc).

Writer is the main scope, but Calc shares the rendering code with
Writer and I don't see a good reason to increase the complexity of
the code by defining different behaviour for Calc.

Change-Id: I26ba02ff02c73a37e8780a67acf4c516c59c193a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194021
Reviewed-by: Tamás Zolnai <zolnaitamas2000@gmail.com>
Tested-by: Jenkins