sc: add dumpAsXml for the pivot table
This adds dumpAsXml for the pivot table, which is triggered with
ctrl+shift+alt F5 when the cursor is inside a pivot table. This
will write the pivot_table_dump.xml into the CWD. The output
includes the layout and the cache, which is similar to the pivot
table dump, that has written the content to the console output.

Change-Id: Ib20089cb63846aafe691a53bdb4e53f16602d5a8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193832
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>