tdf#168917 Sort case-sensitive tooltip improved
It turned out, that there exists no general rule whether uppercase
variants of a letter are treated as being smaller or greater than
lowercase variants. For most latin languages it is 'a' < 'A', for
example. But for Danish it is 'A' < 'a'. Therefore the tooltip is
changed so that it does not mention the order at all.

The fact, that the order is local dependent will be mentioned in the
help.

Change-Id: I6ef39775c8f97ef035b426bd99af49d045ddf9c8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193824
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>