tdf#169219 swpagerelsize ui: immediately apply new RefValue
The m_xWidthED shows either the actual width
or the width as a percentage.
When we switch m_xRelWidthRelationLB,
the base that determines the percentage also needs to change.

Prior to this patch
changing the value from 'Entire Page' to 'Entire Paragraph Area'
did not affect the percentage shown.

Now the percentage increases (only the 'ruler' changed, not the width).

It also means that changing the percentage also correctly
synchronizes the width/height on the Crop tab.

Change-Id: Ib5074fc57972df97c8bf400826e287d0a6070fcb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193831
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins