tdf#162580 copy query text from Find toolbar to Find&Replace dialog
This patch copies the query text to the Find&Replace dialog when Ctrl+H
is pressed in the Find toolbar find text field control. Normally Ctrl+H
toggles the Find&Replace dialog on and off. This patch makes the dialog
not toggle off when pressed in the Find toolbar find text field control,
it either opens the Find&Replace dialog and copies the query text or
copies the query text to the already open Find&Replace dialog.

Change-Id: I6aa28f9e47431bdc75d63c7f68469203bc3b59df
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193837
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>