sc: fix crash in pivot table with formats and no source data
Add more strict checking if the dimension exists as a field so
it can't happend that we crash when we want a dimension that
doesn't exists. Additionally check the group items when filling
the string values of the members, because sometimes a dimension
is a group and is not found as a normal field. Also make sure
that if the dimension is not available we don't crash.

In the test the document has no source data and the "Month" is
a group dimension.

Change-Id: I9db996bb4327b2e081dbfc6dea7d731455111490
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193845
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193998
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>