tdf#62248 fix multiple groupings on same field issue in reportbuilder
In a reportbuilder report, when grouping on the same field multiple times
(with different interval/prefix characters setting), the two groupings interact
and in the end it does not work right; both grouping levels behave like one of
them.

That's because the report function created by Report Builder for each of the
groupings share the same name. Choose unique function name in
ORptExport::exportGroupsExpressionAsFunction and sync OXMLGroup::OXMLGroup to
parse that name. Keep supporting legacy names for backwards compatibility.

Change-Id: Ic91e51743a407225d50b14f85d02c58ca0a243c0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193993
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>