tdf#169219 swrelpagesize ui: value may be changed even if % isn't
Toggling "show percent" off was displaying the previous value
if the percentage hadn't changed.
However, the underlying m_nRefValue might have changed
(like from 'Entire Page' to 'Entire Paragraph')
and so even though the percentage might have remained unchanged (100%)
the width represented by 100% would be different,
so it is not legitimate to just 'show the value from last time'.

The same logic would hold true for percentage with an unchanged width.

The steps-to-reproduce depend on previous patches for this bug report.

-open image properties. Set to 100% relative to Entire Page
-toggle off showing percent: shows something like 4.1 inches
-toggle on percent (100), and change relative to para, set back to 100%
-toggle off showing percent: should show 2.5 inches, not 4...

This has been in the code since initial implementation:
commit 2ecf108c23c041619bb8b25338dca84ff89fe8eb
Author: Caolán McNamara on Thu Dec 20 11:58:29 2012 +0000
    make a PercentFieldWrap that wraps a MetricField

Change-Id: If2e1761cd170ffffb6b99414e717001be0fadd59
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193886
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins