android: Drop two TouchEventHandler members that are always false
As is more obvious since previous commit

    Change-Id: I406a5e4116aaefa673a68257a51a6ae7af8b046c
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Nov 12 16:00:50 2025 +0100

        android: Drop unused TouchEventHandler.setWaitForTouchListeners

, TouchEventHandler.mWaitForTouchListeners is always false,
since that's the default boolean value it gets implicitly
initialized with, and it never gets assigned any other value.

As a consequence, the same is true for
TouchEventHandler.mHoldInQueue as well.

Therefore, drop these two members and all dead code
that would only be executed if one of them were
true.

Change-Id: I145531e23f9d076f79bdb6c0d886f33e05d818f9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193898
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>