tdf#169219 swrelpagesize ui: update on relation changes
Without this patch, when you open up
sw/qa/extras/odfexport/data/relh-page-tdf80282.odt,
go to the image properties,
and change relative from 'Entire Page' to 'Entire Paragraph Area'
and change the updated percent back to 100%,
then when switching to the Crop tab, the width wasn't reduced.
It remained at 4.1 inches.

Now with this patch, the width properly indicates 2.5 inches
and returning to the Position and Size tab remains at 100%.

Change-Id: Ifb8b3b469fb6a70e7c6c7e5953ccccb0e8f1a3b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193908
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>