android: Drop JavaPanZoomController.startingNewEventBlock
... and related logic.

The only (remaining) caller of the method calls it with
a value of false for the `waitingforTouchListeners` param,
in which case the method doesn't do anything.

Also drop documentation related to processing events in
blocks, as that is no longer the case.

TouchEventHandler.isDownEvent is now unused, so
drop it, too.

Change-Id: Ie83eb651500e5beb892492b681e03ce715800069
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193926
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>