android: Drop unused PanZoomState.WAITING_LISTENERS enum value
It's never set, so drop logic only relevant if
it were ever used.

See also previous commit

    Change-Id: Ie83eb651500e5beb892492b681e03ce715800069
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Nov 12 23:09:59 2025 +0100

        android: Drop JavaPanZoomController.startingNewEventBlock

that dropped dead code that would have set the state if
it had ever been reached.

Somewhat related to this, also drop the mention of
"default-prevented" events from the TouchEventHandler doc, as this
is no longer a thing.

Change-Id: I79e918b659099129f733d4fcc7b92d5d35bb6085
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193928
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>