related tdf#162580: copy query text from sidebar Find deck to F&R dialog
Makes the query text in the Find&Replace dialog find entry become that
of the sidebar Find deck find entry when keyboard focus is in the Find
deck find entry and Ctrl+H (or any accelerator assigned to
.uno:SearchDialog) is pressed or the F&R button in the Find deck is
pressed. Normally these actions make the F&R dialog open or close. Here
the F&R dialog is never closed. It either opens the F&R dialog and
copies the find text to it or copies the find text to the already open
F&R dialog. Reasoning for this behavior: F&R dialog is open. Find deck
search finds nothing. Ctrl+H or F&R toolbar button in the Find deck is
pressed to upgrade the search using options in the F&R dialog. Without
this patch the open F&R dialog closes when Ctrl+H or the F&R button is
pressed. Likely what is expected is the F&R dialog remains open and the
text in the F&R dialog find entry becomes that of the sidebar Find deck
find entry, which is what this patch does.

Change-Id: Ic01cbbe6881469349a0756f3837428324f27d5b4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193945
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>