Avoid using local variable that hides a class member
rBHelper was a local variable with the same name of a member variable of
the class ODatabaseMetaDataResultSet. With this patch, we avoid this
name duplication.

The above duplication was causing build problem with the latest VS 2026
compiler:

error C2220: the following warning is treated as an error
warning C4458: declaration of 'rBHelper' hides class member

Change-Id: Id70c690c3b09611aefab720aef1f8c8c1dd46829
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/193965
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>