sc: remove grab bag from pivot table, add PivotTableStyleInfo
This removes the grab bag from the pivot table, which was used to
store the pivotTableStyleInfo element. Instead of this add a struct
to store the style info data and add that to the DPObject. With
this we can actually take the information and use when we output
the pivot table to the sheet.

Change-Id: I1226b80ac0083797c08a20342603c652c43a9568
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194017
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>