UITest: Add a helper function to create an empty database
Adds a method that creates a database document from the start center
button similar to create_doc_in_start_center. Creating a database is a
bit more involved because it first opens a dialog asking for details
about the database. It also needs a filename immediately so a temporary
file is created to store it.

Change-Id: I3e5489ce539be0d9d9148b612b25da828866256f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194040
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins