Add a unit test for selecting between 3 tables when inserting a relation
The test creates a database with three tables and then creates a new
query involving all three tables. It does this so it can open the insert
relation dialog with more than two tables. It then selects various
values for the left table and makes sure that the right table is updated
so that it’s not the same.

Change-Id: Ifd23bb3343a415d828697719c17f5afbff55d856
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194041
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>