Revert "tdf#169219: value may be changed even if % isn't"
This reverts my 26.2 commit 7c2f02497ccc026ccaf48fc3b0b109b78c93c592.
(and the follow-up cleanup commit)

I think I learned the reason for restoring the previous value.

The calculation of the width based on a percentage is less precise
than the decimal-place width value.
In other words, it is more of an approximation.
So, if the percentage was not changed,
then use the precise width stored as the last value,
instead of re-calculating a nearly-identical width.

This means that toggling back and forth will not minutely
change the width value. That precision might be very valuable.

A follow-up commit will add a nLastRefValue,
but with this commit I just want to cleanly
revert back to the original code - without making any changes.

Change-Id: Id940c9a71c8123f6cffbd623f656acfa88ed19a8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194051
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>