tdf#169219 swrelpagesize ui: value may be changed even if % isn't #2
Toggling "show percent" off was displaying the previous value
if the percentage hadn't changed.
However, the underlying m_nRefValue might have changed
(like from 'Entire Page' to 'Entire Paragraph')
and so even though the percentage might have remained unchanged (100%)
the width represented by 100% would be different,
so it is not legitimate to always 'show the value from last time'.

The same logic holds true for percentage with an unchanged width.

The steps-to-reproduce depend on previous patches for this bug report.

-open image properties. Set to 100% relative to Entire Page
-toggle off showing percent: shows something like 4.1 inches
-toggle on percent (100), and change relative to para, set back to 100%
-toggle off showing percent: should show 2.5 inches, not 4...

This has been in the code since initial implementation:
commit 2ecf108c23c041619bb8b25338dca84ff89fe8eb
Author: Caolán McNamara on Thu Dec 20 11:58:29 2012 +0000
make a PercentFieldWrap that wraps a MetricField

Try #1 ignored the fact that using a percentage to calculate the width
is somewhat imprecise, so minor drift could occur.

Change-Id: I1998506e64d8ba247d11772f77d02e565ea5ddcb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194052
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>