lok: sc: use view specific calendar conforming to init. locale/language
The aim is to allow user/view specific locale/language but we are not
there yet. This patch reuses the existing .uno:LanguageStatus to
maintain a view specific locale/language and a calendar based on it. One
problem is that the status-change of this uno command is broadcast to
all views. So if one view sets language X, all client views are forced
to use X. Another issue is that changes in tiles due to one view's
language settings is broadcast to all views. I'm working on fixing these
in a follow-up patches.

Before the patch, when a view selects a language it did not had any
effect on calendar strings like month names in pivot table with date
grouping even after refresh.

Conflicts:
	sc/qa/unit/tiledrendering/tiledrendering.cxx

Signed-off-by: Dennis Francis <dennis.francis@collabora.com>
Change-Id: I92198c8c1467555f27e2642de6c97b4d5e6118b3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194204
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195664
Tested-by: Jenkins