tdf#169413 docx export: don't write themes twice in w:rFonts
A round-trip was reported as corrupt by both Writer and Word.

This probably started in 6.0 with tdf#38778
although a bibisect pointed to my 24.2 patch for tdf#108505.

Although I'm not exactly sure what triggers this,
grabBag was doubling up the bug report's conf12.docx
w:cstheme in rFonts which looked like
    <w:rFonts w:cstheme="minorHAnsi"
              w:cs="Calibri"
              w:cstheme="minorHAnsi"/>
and duplicate entries are invalid XML - even firefox complains.

Change-Id: Id9561b940c2c536147422e10a8c1d37ea51c2e78
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194057
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194064
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>