ScriptForge Timer uses Basic Now function
The durations in the Timer service
are measured with a precision of
the millisecond.

Thanks to tdf#161469 the Basic Now()
function provides sub-seconds precision
since LO 25.2.

Hence the use of the Calc NOW() function
is replaced in the Basic code by the
more immediately available Basic Now()
equivalent.

No impact on existing user scripts.
Documentation is unchanged.

Change-Id: Ia3b0f07918d9b9a9e96f62aaa5ce3862a2b48579
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194070
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>