tdf#130857 Define widgets in Borders tab page in order
Switch the order in which the widgets in the .ui file
are defined so that the order better matches the visual appearance
and logical order, which makes sure that tab focus order with the
Qt-based VCL plugins is correct as well when using native Qt
widgets (SAL_VCL_QT_USE_WELDED_WIDGETS=1), where
focus order is (by default) based on widget creation order.

See

    commit 02692566ad9fc7c3484f8581ffa0004cd4e43987
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 24 17:43:35 2024 +0200

        tdf#130857 optnewdictionarydialog.ui: Define focusable widgets in order

for more background.

Using native Qt widgets isn't yet supported for that
tab page, but this makes focus order work as expected
in a WIP branch to implement that.

Sample steps to trigger the "Borders" dialog using
that tab page:

* start Writer
* "Tools" -> "Customize"
* add "Borders" entry to the menu and close the dialog
* open menu and click the newly added menu entry

Change-Id: Iae68d1f4d09f52dc68ddc487d62ae03b24c596db
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194076
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>