cool#13468 sw markdown paste: add UNO command parameter to skip the detection
.uno:Paste tries to detect if the plain text content is really just
plain text or markdown. There is no way to paste the clipboard content
as plain text in a non-interactive way if the input looks like markdown.

What happens is that .uno:Paste uses comphelper::IsMarkdownData() to
guess if the input is markdown. This is fine in general, but if you want
to paste "foo _bar_ baz" as-is, there is no .uno:Paste parameter to do
so, you always get "foo <i>bar</i> baz".

Fix the problem by adding a new SkipDetection parameter for .uno:Paste,
while keeping the current behavior unchanged. The LOK API's doc_paste()
function has an explicit mimetype parameter, so skip detection when
using that API (a LOK client will use doc_setClipboard() when the user
does a Ctrl-V, behavior there is unchanged).

Note that an interactive way did exist previously: if you did paste
special and selected plain text, that did work.

Change-Id: I9888b1bf2322d3e1eeb32ffc8c8d08490b8eaa2c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194110
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>